chore: moved usertimetable hook to contexts
